% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_bootstrap.R
\name{coef.fpgs_bootstrap}
\alias{coef.fpgs_bootstrap}
\title{Extract FPGS Bootstrap Estimate}
\usage{
\method{coef}{fpgs_bootstrap}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fpgs_bootstrap"}.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
The estimated average treatment effect.
}
\description{
Extracts the estimated average treatment effect from an FPGS
bootstrap object.
}
