% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgs_bootstrap.R
\name{vcov.fpgs_bootstrap}
\alias{vcov.fpgs_bootstrap}
\title{Variance-Covariance Matrix for FPGS Bootstrap Estimate}
\usage{
\method{vcov}{fpgs_bootstrap}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fpgs_bootstrap"}.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
A one-by-one matrix containing the bootstrap variance estimate.
}
\description{
Returns the bootstrap variance estimate as a one-by-one
variance-covariance matrix.
}
